+++ gazebo/physics/Base.hh
@@ -15,13 +15,8 @@
  *
 */
 
-/* Desc: Base class shared by all classes in Gazebo.
- * Author: Nate Koenig
- * Date: 09 Sept. 2008
- */
-
-#ifndef _GAZEBO_PHYSICS_BASE_HH_
-#define _GAZEBO_PHYSICS_BASE_HH_
+#ifndef GAZEBO_PHYSICS_BASE_HH_
+#define GAZEBO_PHYSICS_BASE_HH_
 
 #ifdef _WIN32
   // Ensure that Winsock2.h is included before Windows.h, which can get
@@ -35,6 +30,7 @@
 #include <sdf/sdf.hh>
 
 #include "gazebo/common/CommonTypes.hh"
+#include "gazebo/common/URI.hh"
 #include "gazebo/physics/PhysicsTypes.hh"
 #include "gazebo/util/system.hh"
 
@@ -93,8 +89,6 @@ namespace gazebo
                 LINK            = 0x00000004,
                 /// \brief Collision type
                 COLLISION       = 0x00000008,
-                /// \brief Actor type
-                ACTOR           = 0x00000016,
                 /// \brief Light type
                 LIGHT           = 0x00000010,
                 /// \brief Visual type
@@ -118,6 +112,9 @@ namespace gazebo
                 GEARBOX_JOINT   = 0x00002000,
                 /// \brief FixedJoint type
                 FIXED_JOINT     = 0x00004000,
+
+                /// \brief Actor type
+                ACTOR           = 0x00008000,
 
                 /// \brief Shape type
                 SHAPE           = 0x00010000,
@@ -269,6 +266,10 @@ namespace gazebo
       /// \return The full type definition.
       public: unsigned int GetType() const;
 
+      /// \brief Get the string name for the entity type.
+      /// \return The string name for this entity.
+      public: std::string TypeStr() const;
+
       /// \brief Return the name of this entity with the model scope
       /// model1::...::modelN::entityName
       /// \param[in] _prependWorldName True to prended the returned string
@@ -276,6 +277,14 @@ namespace gazebo
       /// world::model1::...::modelN::entityName.
       /// \return The scoped name.
       public: std::string GetScopedName(bool _prependWorldName = false) const;
+
+      /// \brief Return the common::URI of this entity.
+      /// The URI includes the world where the entity is contained and all the
+      /// hierarchy of sub-entities that can compose this entity.
+      /// E.g.: A link entity contains the name of the link and the model where
+      /// the link is contained.
+      /// \return The URI of this entity.
+      public: common::URI URI() const;
 
       /// \brief Print this object to screen via gzmsg.
       /// \param[in] _prefix Usually a set of spaces.
@@ -334,6 +343,9 @@ namespace gazebo
 
       /// \brief The type of this object.
       private: unsigned int type;
+
+      /// \brief The string representation of the type of this object.
+      private: std::string typeStr;
 
       /// \brief True if selected.
       private: bool selected;