+++ gazebo/physics/Link.cc
@@ -47,6 +47,7 @@
 #include "gazebo/physics/PhysicsEngine.hh"
 #include "gazebo/physics/Collision.hh"
 #include "gazebo/physics/Link.hh"
+#include "gazebo/physics/Wind.hh"
 
 using namespace gazebo;
 using namespace physics;
@@ -187,6 +188,17 @@ void Link::Load(sdf::ElementPtr _sdf)
     }
   }
 
+  if (this->sdf->HasElement("enable_wind"))
+  {
+    this->SetWindMode(this->sdf->Get<bool>("enable_wind"));
+  }
+  else
+  {
+    this->SetWindMode(this->GetModel()->WindMode());
+  }
+  this->sdf->GetElement("enable_wind")->GetValue()->SetUpdateFunc(
+      std::bind(&Link::WindMode, this));
+
   this->connections.push_back(event::Events::ConnectWorldUpdateBegin(
       boost::bind(&Link::Update, this, _1)));
 
@@ -225,6 +237,9 @@ void Link::Init()
     battery->Init();
   }
 
+  if (this->WindMode() && this->world->WindEnabled())
+    this->SetWindEnabled(true);
+
   this->initialized = true;
 }
 
@@ -255,6 +270,7 @@ void Link::Fini()
       auto deleteMsg = msgs::CreateRequest("entity_delete",
           std::to_string(iter.second.id()));
       this->requestPub->Publish(*deleteMsg, true);
+      delete deleteMsg;
 
       msgs::Visual msg;
       msg.set_name(iter.second.name());
@@ -336,11 +352,15 @@ void Link::UpdateParameters(sdf::ElementPtr _sdf)
 
   this->sdf->GetElement("gravity")->GetValue()->SetUpdateFunc(
       boost::bind(&Link::GetGravityMode, this));
+  this->sdf->GetElement("enable_wind")->GetValue()->SetUpdateFunc(
+      std::bind(&Link::WindMode, this));
   this->sdf->GetElement("kinematic")->GetValue()->SetUpdateFunc(
       boost::bind(&Link::GetKinematic, this));
 
   if (this->sdf->Get<bool>("gravity") != this->GetGravityMode())
     this->SetGravityMode(this->sdf->Get<bool>("gravity"));
+
+  this->SetWindMode(this->sdf->Get<bool>("enable_wind"));
 
   // before loading child collision, we have to figure out if
   // selfCollide is true and modify parent class Entity so this
@@ -511,6 +531,12 @@ void Link::Update(const common::UpdateInfo & /*_info*/)
   {
     battery->Update();
   }
+}
+
+//////////////////////////////////////////////////
+void Link::UpdateWind(const common::UpdateInfo & /*_info*/)
+{
+  this->windLinearVel = this->world->Wind().WorldLinearVel(this);
 }
 
 /////////////////////////////////////////////////
@@ -722,6 +748,46 @@ math::Box Link::GetBoundingBox() const
 }
 
 //////////////////////////////////////////////////
+void Link::SetWindMode(const bool _mode)
+{
+  this->sdf->GetElement("enable_wind")->Set(_mode);
+
+  if (!this->WindMode() && this->updateConnection)
+    this->SetWindEnabled(false);
+  else if (this->WindMode() && !this->updateConnection)
+    this->SetWindEnabled(true);
+}
+
+/////////////////////////////////////////////////
+void Link::SetWindEnabled(const bool _enable)
+{
+  if (_enable)
+  {
+    this->updateConnection = event::Events::ConnectWorldUpdateBegin(
+        std::bind(&Link::UpdateWind, this, std::placeholders::_1));
+  }
+  else
+  {
+    event::Events::DisconnectWorldUpdateBegin(this->updateConnection);
+    this->updateConnection.reset();
+    // Make sure wind velocity is null
+    this->windLinearVel.Set(0, 0, 0);
+  }
+}
+
+//////////////////////////////////////////////////
+const ignition::math::Vector3d Link::WorldWindLinearVel() const
+{
+  return this->windLinearVel;
+}
+
+//////////////////////////////////////////////////
+bool Link::WindMode() const
+{
+  return this->sdf->Get<bool>("enable_wind");
+}
+
+//////////////////////////////////////////////////
 bool Link::SetSelected(bool _s)
 {
   Entity::SetSelected(_s);
@@ -813,6 +879,7 @@ void Link::FillMsg(msgs::Link &_msg)
   _msg.set_name(this->GetScopedName());
   _msg.set_self_collide(this->GetSelfCollide());
   _msg.set_gravity(this->GetGravityMode());
+  _msg.set_enable_wind(this->WindMode());
   _msg.set_kinematic(this->GetKinematic());
   _msg.set_enabled(this->GetEnabled());
   msgs::Set(_msg.mutable_pose(), relPose.Ign());
@@ -908,6 +975,8 @@ void Link::ProcessMsg(const msgs::Link &_msg)
     this->SetGravityMode(_msg.gravity());
     this->SetEnabled(true);
   }
+  if (_msg.has_enable_wind())
+    this->SetWindMode(_msg.enable_wind());
   if (_msg.has_kinematic())
   {
     this->SetKinematic(_msg.kinematic());
@@ -1612,4 +1681,11 @@ void Link::LoadBattery(sdf::ElementPtr _sdf)
   common::BatteryPtr battery(new common::Battery());
   battery->Load(_sdf);
   this->batteries.push_back(battery);
+}
+
+/////////////////////////////////////////////////
+const ignition::math::Vector3d Link::RelativeWindLinearVel() const
+{
+  return this->GetWorldPose().Ign().Rot().Inverse().RotateVector(
+      this->windLinearVel);
 }