+++ gazebo/physics/ode/ODESliderJoint.cc
@@ -58,7 +58,10 @@ math::Vector3 ODESliderJoint::GetGlobalAxis(unsigned int /*_index*/) const
   if (this->jointId)
     dJointGetSliderAxis(this->jointId, result);
   else
+  {
     gzerr << "ODE Joint ID is invalid\n";
+    return math::Vector3::Zero;
+  }
 
   return math::Vector3(result[0], result[1], result[2]);
 }
@@ -146,9 +149,8 @@ double ODESliderJoint::GetParam(unsigned int _parameter) const
 //////////////////////////////////////////////////
 math::Vector3 ODESliderJoint::GetAnchor(unsigned int /*_index*/) const
 {
-  dVector3 result;
   gzlog << "ODESliderJoint::GetAnchor not implemented.\n";
-  return math::Vector3(result[0], result[1], result[2]);
+  return math::Vector3::Zero;
 }
 
 //////////////////////////////////////////////////