+++ tools/gz_log.hh
@@ -31,6 +31,7 @@ namespace gazebo
     /// \brief Constructor
     /// \param[in] _xmlOutput True if the output should be in XML format.
     /// \param[in[ _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     public: FilterBase(bool _xmlOutput, const std::string &_stamp);
 
     /// \brief Output a line of data.
@@ -64,6 +65,7 @@ namespace gazebo
     /// \brief Constructor.
     /// \param[in] _xmlOutput True if the output should be in XML format.
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     public: JointFilter(bool _xmlOutput, const std::string &_stamp);
 
     /// \brief Initialize the filter.
@@ -93,6 +95,7 @@ namespace gazebo
     /// \brief Constructor.
     /// \param[in] _xmlOutput True if the output should be in XML format.
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     public: LinkFilter(bool _xmlOutput, const std::string &_stamp);
 
     /// \brief Initialize the filter.
@@ -122,6 +125,7 @@ namespace gazebo
     /// \brief Constructor.
     /// \param[in] _xmlOutput True if the output should be in XML format.
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     public: ModelFilter(bool _xmlOutput, const std::string &_stamp);
 
     /// \brief Destructor.
@@ -160,6 +164,7 @@ namespace gazebo
     /// \brief Constructor
     /// \param[in] _xmlOutput True to format output as XML
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     public: StateFilter(bool _xmlOutput, const std::string &_stamp,
                 double _hz = 0);
 
@@ -201,10 +206,27 @@ namespace gazebo
     /// \param[in] _filename Name of the file to parse.
     private: void Info(const std::string &_filename);
 
+    /// \brief Output log data to a file. This is usually used with the
+    /// filter command.
+    /// \param[in] _outFilename Output filename
+    /// \param[in] _filter Filter string
+    /// \param[in] _raw True to output data without xml formatting.
+    /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
+    /// \param[in] _hz Hertz rate.
+    /// \param[in] _encoding Specify output log file encoding. If empty, the
+    /// encoding from the source log file is used.
+    /// Valid values include (txt, zlib, bz2)
+    private: void Output(const std::string &_outFilename,
+                 const std::string &_filter, const bool _raw,
+                 const std::string &_stamp, const double _hz,
+                 const std::string &_encoding = "");
+
     /// \brief Dump the contents of a log file to screen
     /// \param[in] _filter Filter string
     /// \param[in] _raw True to output data without xml formatting.
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     /// \param[in] _hz Hertz rate.
     private: void Echo(const std::string &_filter,
                  bool _raw, const std::string &_stamp, double _hz);
@@ -213,6 +235,7 @@ namespace gazebo
     /// \param[in] _filter Filter string
     /// \param[in] _raw True to output data without xml formatting.
     /// \param[in] _stamp Type of stamp to apply.
+    /// Valid values are (sim,real,wall)
     /// \param[in] _hz Hertz rate.
     private: void Step(const std::string &_filter, bool _raw,
                  const std::string &_stamp, double _hz);
@@ -235,6 +258,16 @@ namespace gazebo
     /// \param[in] _filename Filename to open
     /// \return True on success.
     private: bool LoadLogFromFile(const std::string &_filename);
+
+
+    /// \brief Write data to a file.
+    /// \param[in] _outFile Output file stream reference.
+    /// \param[in] _stateString SDF state string to write
+    /// \param[in] _raw True to output data without xml formatting.
+    /// \param[in] _encoding Encoding type: txt, zlib, bz2
+    private: void OutputWriter(std::ofstream &_outFile,
+                 const std::string &_stateString,
+                 const bool _raw, const std::string &_encoding);
 
     /// \brief Node pointer.
     private: gazebo::transport::NodePtr node;